Moves long chunks of template code to "snippets"
Sometimes we want to use the default template, only modified a little
bit (as opposed to using a completely custom template). This way we can
probably take any updates easily, but still add a little custom
behavior. In these cases, we want to specify our template modifications
outside of the main application template file, so we put them in their
own files in the snippets folder (that folder’s structure mimics the
Rails application - mostly).